+++ src/Components/IssueHeader.tsx
@@ -5,6 +5,8 @@ import Button from '@material-ui/core/Button';
 import ButtonGroup from '@material-ui/core/ButtonGroup';
 import Hidden from '@material-ui/core/Hidden';
 import Grid from '@material-ui/core/Grid';
+import { issueStore } from '../index';
+import { Link as RouterLink } from 'react-router-dom';
 
 /**
  * Displays a user board with issue cards to the users, and buttons to create and filter issues
@@ -22,18 +24,23 @@ const useStyles = makeStyles((theme: Theme) => ({
 const IssueHeader = () => {
   const classes = useStyles();
 
+  //updates the issue store's issueFilter
+  const handleApplyFilter = (filter: string) => {
+    issueStore.setIssueFilter(filter);
+  };
+
   return (
     <Grid container className={classes.root}>
       <Grid item className={classes.buttonGroup}>
         <ButtonGroup color="default" variant="contained">
-          <Button>All Issues</Button>
-          <Button>Unresolved Issues</Button>
-          <Button>Urgent Issues</Button>
+          <Button onClick={() => handleApplyFilter('')}>All Issues</Button>
+          <Button onClick={() => handleApplyFilter('unresolved')}>Unresolved Issues</Button>
+          <Button onClick={() => handleApplyFilter('urgent')}>Urgent Issues</Button>
         </ButtonGroup>
       </Grid>
       <Hidden xsDown>
         <Grid item>
-          <Button variant="contained" color="primary">
+          <Button variant="contained" color="primary" component={RouterLink} to="/new-issue">
             New Issue
           </Button>
         </Grid>